Rename avsEigenMRP to eigenMRP
