+++ src/fswAlgorithms/imageProcessing/centerRadiusCNN/centerRadiusCNN.h
@@ -32,7 +32,7 @@
 #include "architecture/msgPayloadDef/OpNavCirclesMsgPayload.h"
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 
 /*! @brief The CNN based center radius visual tracking module. */
