+++ src/fswAlgorithms/imageProcessing/houghCircles/houghCircles.h
@@ -34,7 +34,7 @@
 #include "architecture/msgPayloadDef/OpNavCirclesMsgPayload.h"
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 
 /*! @brief visual planet tracking with Hough circles */
