+++ src/fswAlgorithms/imageProcessing/opticalFlow/opticalFlow.h
@@ -34,7 +34,7 @@
 #include "architecture/msgPayloadDef/PairedKeyPointsMsgPayload.h"
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/bskLogging.h"
 #include "architecture/utilities/linearAlgebra.h"
