+++ src/fswAlgorithms/pointCloudProcessing/SICP/scalingIterativeClosestPoint.h
@@ -29,7 +29,7 @@
 #include "architecture/msgPayloadDef/SICPMsgPayload.h"
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/bskLogging.h"
