+++ src/simulation/dynamics/DynOutput/boreAngCalc/boreAngCalc.h
@@ -28,7 +28,7 @@
 #include "architecture/msgPayloadDef/BoreAngleMsgPayload.h"
 #include "architecture/msgPayloadDef/SCStatesMsgPayload.h"
 #include "architecture/msgPayloadDef/SpicePlanetStateMsgPayload.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/bskLogging.h"
