+++ src/simulation/dynamics/HingedRigidBodies/hingedRigidBodyStateEffector.h
@@ -21,7 +21,7 @@
 #define HINGED_RIGID_BODY_STATE_EFFECTOR_H
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "simulation/dynamics/_GeneralModuleFiles/stateData.h"
 #include "simulation/dynamics/_GeneralModuleFiles/stateEffector.h"
 #include <Eigen/Dense>
