+++ src/simulation/dynamics/facetDragEffector/facetDragDynamicEffector.cpp
@@ -20,7 +20,7 @@
 #include "facetDragDynamicEffector.h"
 
 #include "architecture/utilities/astroConstants.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
