+++ src/simulation/dynamics/prescribedMotion/prescribedMotionStateEffector.h
@@ -25,7 +25,7 @@
 #include "architecture/msgPayloadDef/PrescribedRotationMsgPayload.h"
 #include "architecture/msgPayloadDef/PrescribedTranslationMsgPayload.h"
 #include "architecture/msgPayloadDef/SCStatesMsgPayload.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "simulation/dynamics/_GeneralModuleFiles/stateData.h"
 #include "simulation/dynamics/_GeneralModuleFiles/stateEffector.h"
