+++ src/simulation/dynamics/reactionWheels/reactionWheelStateEffector.h
@@ -37,7 +37,7 @@
 #include "architecture/messaging/messaging.h"
 #include "architecture/utilities/bskLogging.h"
 
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 
 /*! @brief reaction wheel state effector class */
