+++ src/simulation/dynamics/spacecraft/spacecraft.cpp
@@ -19,7 +19,7 @@
 
 #include "spacecraft.h"
 #include "../_GeneralModuleFiles/svIntegratorRK4.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/macroDefinitions.h"
 #include <iostream>
