+++ src/simulation/dynamics/stateSpaceEffector/stateSpaceStateEffector.h
@@ -24,7 +24,7 @@
 #include "simulation/dynamics/_GeneralModuleFiles/stateEffector.h"
 #include "simulation/dynamics/_GeneralModuleFiles/stateData.h"
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 
 class StateSpaceStateEffector: public StateEffector, public SysModel {
