+++ src/simulation/environment/albedo/albedo.h
@@ -21,7 +21,7 @@
 #define ALBEDO_H
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 
 #include "architecture/messaging/messaging.h"
