+++ src/simulation/navigation/planetHeading/planetHeading.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 #include <Eigen/Dense>
