+++ src/simulation/power/simpleSolarPanel/simpleSolarPanel.cpp
@@ -3,7 +3,7 @@
 //
 #include "simpleSolarPanel.h"
 #include "architecture/utilities/astroConstants.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
 #include <math.h>
