+++ src/simulation/sensors/camera/camera.h
@@ -36,7 +36,7 @@
 #include "architecture/msgPayloadDef/CameraModelMsgPayload.h"
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 
 /*! @brief visual camera class */
