+++ src/simulation/sensors/coarseSunSensor/coarseSunSensor.cpp
@@ -19,7 +19,7 @@
 
 #include "simulation/sensors/coarseSunSensor/coarseSunSensor.h"
 #include "architecture/utilities/astroConstants.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/macroDefinitions.h"
