+++ src/simulation/sensors/imuSensor/imuSensor.h
@@ -31,7 +31,7 @@
 #include "architecture/msgPayloadDef/IMUSensorMsgPayload.h"
 #include "architecture/msgPayloadDef/SCStatesMsgPayload.h"
 
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/bskLogging.h"
 #include "architecture/utilities/macroDefinitions.h"
 #include <Eigen/Dense>
