+++ src/simulation/thermal/sensorThermal/sensorThermal.cpp
@@ -1,5 +1,5 @@
 #include "sensorThermal.h"
-#include "architecture/utilities/avsEigenMRP.h"
+#include "architecture/utilities/eigenMRP.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/macroDefinitions.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
